# server/openapi.yaml
openapi: 3.0.3
info:
  title: LEGION
  version: 0.1.0
  description: The engine. One Django server, every limb a client of it.
paths:
...
  /api/events:
    get:
      operationId: api_events_list
      description: '`GET /api/events?since=<iso>` and `POST /api/events`.'
      parameters:
      - in: query
        name: since
        schema:
          type: string
          format: date-time
        description: ISO-8601 UTC watermark. Returns rows with `updated_at >= since`,
          tombstones included, oldest first. **Omitted or unparsable means EVERYTHING,
          never nothing** (api/sync.parse_since), so a fresh client's first pull is
          the whole table. Hand back the `next` cursor from the previous page, or
          `server_time` from GET /api/changes; both are already `Z`-suffixed so they
          need no extra encoding.
      tags:
      - events
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEvent'
          description: Rows changed at or after `since`, tombstones included, oldest
            first, 500 to a page. This route does NOT accept `?active=1` - the synced
            routes do; here a client filters `deleted_at` itself.
    post:
      operationId: api_events_create
      description: '`GET /api/events?since=<iso>` and `POST /api/events`.'
      tags:
      - events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Event'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      security:
      - deviceToken: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: Created. The body is the row as stored, server-minted `id`,
            `provenance` and timestamps included.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: '**Not created - this `origin_guid` already exists**, and the
            body is the row that was already there, unchanged. A retried create is
            a no-op, which is what makes a lost acknowledgement safe to retry. A client
            tells the two apart by the status code, never by the body.'
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: 'The write was refused and NOTHING was written. Two body shapes
            occur and both are JSON objects: `{"detail": "..."}` for an unknown field,
            a database refusal or a hand-written check, and `{"<field>": ["..."]}`
            for a field-level validation error. Both carry text meant to be shown
            to a person.'
  /api/events/{id}:
    patch:
      operationId: api_events_partial_update
      description: '`PATCH`/`DELETE /api/events/<id>`.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedEvent'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedEvent'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedEvent'
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: The row as stored. Every field is optional on the way in; only
            the ones sent are changed. Sending `done` without `done_at` derives `done_at`
            from the database clock (or clears it when `done` is false).
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: 'The write was refused and NOTHING was written. Two body shapes
            occur and both are JSON objects: `{"detail": "..."}` for an unknown field,
            a database refusal or a hand-written check, and `{"<field>": ["..."]}`
            for a field-level validation error. Both carry text meant to be shown
            to a person.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: No such row. Nothing was changed.
    delete:
      operationId: api_events_destroy
      description: '`PATCH`/`DELETE /api/events/<id>`.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - events
      security:
      - deviceToken: []
      responses:
        '204':
          description: Done. Soft-deleted (the tombstone row stays so an unsynced
            client learns of it).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: No such row. Nothing was changed.
...
components:
  schemas:
...
    Detail:
      type: object
      description: |-
        `{"detail": "..."}` - the shape every hand-written refusal in this API
        uses, and the shape DRF itself uses for 401/403/404/405. The strings are
        written to be shown to a person: they say what did NOT happen (CLAUDE.md
        section 7's outcome-verb rule), so a client should surface `detail`
        rather than inventing its own wording.
      properties:
        detail:
          type: string
      required:
      - detail
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        all_day:
          type: boolean
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        source:
          type: string
        google_event_id:
          type: string
          nullable: true
        done:
          type: boolean
        done_at:
          type: string
          format: date-time
          nullable: true
        sort_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        trigger_place_label:
          type: string
          nullable: true
        repeat_kind:
          type: string
          nullable: true
        repeat_every:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_days_of_week:
          type: string
          nullable: true
        repeat_day:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_month:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_end_kind:
          type: string
          nullable: true
        repeat_end_date:
          type: string
          format: date
          nullable: true
        repeat_end_count:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        exact:
          type: boolean
        exact_downgraded:
          type: boolean
        missed_at:
          type: string
          format: date-time
          nullable: true
        missed_dismissed_at:
          type: string
          format: date-time
          nullable: true
        logged_at:
          type: string
          format: date-time
          nullable: true
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
          nullable: true
        structured_meta:
          nullable: true
        kind:
          type: string
      required:
      - created_at
      - deleted_at
      - id
      - provenance
      - title
      - updated_at
...
    PagedEvent:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next:
          type: string
          format: date-time
          nullable: true
          description: 'Cursor for the next page: hand it back as `?since=`. Null
            means this was the last page.'
      required:
      - next
      - results
...
    PatchedEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        all_day:
          type: boolean
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        source:
          type: string
        google_event_id:
          type: string
          nullable: true
        done:
          type: boolean
        done_at:
          type: string
          format: date-time
          nullable: true
        sort_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        trigger_place_label:
          type: string
          nullable: true
        repeat_kind:
          type: string
          nullable: true
        repeat_every:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_days_of_week:
          type: string
          nullable: true
        repeat_day:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_month:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_end_kind:
          type: string
          nullable: true
        repeat_end_date:
          type: string
          format: date
          nullable: true
        repeat_end_count:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        exact:
          type: boolean
        exact_downgraded:
          type: boolean
        missed_at:
          type: string
          format: date-time
          nullable: true
        missed_dismissed_at:
          type: string
          format: date-time
          nullable: true
        logged_at:
          type: string
          format: date-time
          nullable: true
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
          nullable: true
        structured_meta:
          nullable: true
        kind:
          type: string
...
    ProvenanceEnum:
      enum:
      - DETERMINISTIC
      - LLM_RECONCILED
      - UNRECONCILED
      - USER
      type: string
      description: |-
        * `DETERMINISTIC` - Deterministic
        * `LLM_RECONCILED` - Llm Reconciled
        * `UNRECONCILED` - Unreconciled
        * `USER` - User
...
  securitySchemes:
    deviceToken:
      type: apiKey
      in: header
      name: Authorization
      description: One token per device (phone, browser, future robot), revocable
        alone. Send the value as `Token <key>`, including the word Token and the space.
        Get a key from POST /api/auth/login; POST /api/auth/logout revokes only the
        key that made the request.